combine input states into input.py, remove some custom input number attributes
