--- src/hassette/models/states/input_text.py
@@ -1,18 +1,0 @@
-from typing import Any, Literal
-
-from pydantic import Field
-
-from .base import AttributesBase, StringBaseState
-
-
-class InputTextState(StringBaseState):
-    class Attributes(AttributesBase):
-        editable: bool | None = Field(default=None)
-        min: int | float | None = Field(default=None)
-        max: int | float | None = Field(default=None)
-        pattern: Any | None = Field(default=None)
-        mode: str | None = Field(default=None)
-
-    domain: Literal["input_text"]
-
-    attributes: Attributes
